git fix

+++ src/main/java/c4/corpsecomplex/core/inventory/DeathStackHelper.java
@@ -9,20 +9,11 @@ import net.minecraftforge.energy.IEnergyStorage;
 
 import java.util.Random;
 
-public class DeathStackHandler {
+public final class DeathStackHelper {
 
-    public boolean isCursed;
-    public boolean isEssential;
-    public boolean store;
-    public ItemStack stack;
     private static Random generator = new Random();
 
-    public DeathStackHandler(ItemStack stack) {
-        this.stack = stack;
-        isCursed = isCursed();
-        isEssential = !isCursed && isEssential();
-        store = false;
-    }
+    private DeathStackHelper() {}
 
     public static ItemStack stackToStore(EntityPlayer player, ItemStack stack, boolean cfgStore) {
 
@@ -30,31 +21,42 @@ public class DeathStackHandler {
         boolean essential = !cursed && isEssential(stack);
         boolean store = ((cfgStore && !cursed) || essential);
 
-        if (cursed) {
+        if (cursed && InventoryModule.destroyCursed) {
             destroyCursed(stack);
-            return ItemStack.EMPTY;
+            return stack;
         }
 
-        loseDurability(player, stack, store);
-        loseEnergy(stack, store);
+        if (InventoryModule.dropLoss > 0 || InventoryModule.keptLoss > 0) {
+            loseDurability(player, stack, store);
+        }
 
-        int keepAmount = stack.getCount();
-        keepAmount -= randomlyDrop(stack);
+        if (InventoryModule.dropDrain > 0 || InventoryModule.keptDrain > 0) {
+            loseEnergy(stack, store);
+        }
 
-        return stack.splitStack(keepAmount);
+        if (stack.isEmpty()) { return ItemStack.EMPTY; }
+
+        if (store) {
+            if (!essential && InventoryModule.randomDrop > 0) {
+                int keepAmount = stack.getCount();
+                keepAmount -= randomlyDrop(stack);
+                return stack.splitStack(keepAmount);
+            } else {
+                ItemStack stack1 = stack.copy();
+                stack.setCount(0);
+                return stack1;
+            }
+        } else {
+            return ItemStack.EMPTY;
+        }
     }
 
-    public static boolean destroyCursed(ItemStack stack) {
-
-        if (!InventoryModule.destroyCursed){ return false; }
+    public static void destroyCursed(ItemStack stack) {
 
         stack.setCount(0);
-        return true;
     }
 
     public static int randomlyDrop(ItemStack stack) {
-
-        if (InventoryModule.randomDrop == 0) { return 0; }
 
         int dropAmount = 0;
         for (int i = 0; i < stack.getCount(); i++) {
@@ -68,8 +70,6 @@ public class DeathStackHandler {
 
     public static void randomlyDestroy(ItemStack stack) {
 
-        if (InventoryModule.randomDestroy == 0) { return; }
-
         int destroyAmount = 0;
         for (int i = 0; i < stack.getCount(); i++) {
             if (generator.nextDouble() < InventoryModule.randomDestroy) {
@@ -82,7 +82,7 @@ public class DeathStackHandler {
 
     public static void loseDurability (EntityPlayer player, ItemStack stack, boolean store) {
 
-        if (!InventoryModule.enableDurability || !stack.isItemStackDamageable()) { return;}
+        if (!stack.isItemStackDamageable()) { return;}
 
         if (store) {
             stack.damageItem((int) Math.round(stack.getMaxDamage()* InventoryModule.keptLoss), player);
@@ -95,7 +95,7 @@ public class DeathStackHandler {
 
         IEnergyStorage energy = stack.getCapability(CapabilityEnergy.ENERGY, null);
 
-        if (energy == null || !InventoryModule.enableEnergy) { return;}
+        if (energy == null) { return;}
 
         int energyToLose;
 
@@ -110,7 +110,7 @@ public class DeathStackHandler {
         }
     }
 
-    private static boolean isEssential(ItemStack stack) {
+    public static boolean isEssential(ItemStack stack) {
 
         for (String s : InventoryModule.essentialItems) {
 
@@ -122,10 +122,11 @@ public class DeathStackHandler {
                 return true;
             }
         }
+
         return false;
     }
 
-    private static boolean isCursed(ItemStack stack) {
+    public static boolean isCursed(ItemStack stack) {
 
         for (String s : InventoryModule.cursedItems) {
 
@@ -137,6 +138,7 @@ public class DeathStackHandler {
                 return true;
             }
         }
+
         return false;
     }
 }
